refactor: extract direct answer recovery
